Extract AbstractPost extension to it's own file
